backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-cors-headers==3.13.0
django-filter==23.1
django-link-shortener==0.5
djoser==2.1.0

gunicorn==20.1.0 

flake8==6.0.0
flake8-isort==6.0.0
flake8-quotes==3.4.0

psycopg2-binary==2.9.3
Pillow==9.0.0
pdfkit==1.0.0